requirements.txt
# Core packages
requests==2.31.0
python-dotenv==1.0.0
watchdog==6.0.0

# Development tools
pytest==7.4.3
black==23.12.1
flake8==6.1.0
mypy==1.8.0

# Data science (optional - uncomment if needed)
# numpy==1.24.3
# pandas==2.0.3
# matplotlib==3.7.2 